Customize mode new assignment syntax, part 1
